사칙연산 추가

+++ 1_단계별로_풀어보기/1_입출력과사칙연산/5_A÷B/5_A÷B.cpp
@@ -4,9 +4,9 @@ using namespace std;
 
 int main()
 {
-    int a;
-    int b;
+    double a;
+    double b;
     cin >> a >> b;
-    cout << (float)a / (float)b;
+    cout << a / b;
     return 0;
 }
